requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
openai-whisper>=20230918
numpy>=1.20.0
pydub>=0.25.1
librosa>=0.9.2
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
requests>=2.28.0
ffmpeg-python>=0.2.0